requirements.txt
# Core data processing and analysis
numpy>=1.24.3
pandas>=2.0.3
pandas-ta>=0.3.14b0  # Technical analysis library

# Data fetching and API interactions
yfinance==0.2.28  # Yahoo Finance API (older version without curl_cffi dependency)
requests>=2.31.0
websocket-client>=1.6.1

# Machine Learning and AI
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.1
transformers>=4.31.0

# Backtesting and Strategy Development
backtrader>=1.9.78.123
vectorbt>=0.26.0  # Vectorized backtesting
numba>=0.57.1  # Required by vectorbt

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Database
sqlalchemy>=2.0.20
pymongo>=4.5.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking and linting
mypy>=1.5.1
black>=23.7.0
flake8>=6.1.0 